Rewrite `remote:copy()': handle pulls and pushes
Up until now, the only thing this function was capable of doing was to
pull a file from a remote host, with the bonus being that it handled
ssh bounce proxies based on the environment config.

It can now do remote to local, local to remote, remote to remote, and
local to local.

The remote to remote requires further work however, might be an option
to allow the local host to act as an intermediate where the two remote
hosts don't have access to one-another - which is often the case in
many environments (for example, development to production).