Feature: module groups; closes schtunt/simbol#10
Note that the library code does not yet support this, but that will
come in before further work on this module (being the first to utilize
this new feature) commences.

With this new feature, you can group related modules into a
subdirectory, for example:

    aws/ec2.sh
    aws/s3.sh
    aws/sns.sh
    aws/sqs.sh

This makes sense for modules with a very large potential to grow very
large, like AWS.

Nesting can only be 1 level deep.