Fix bash completion
The `simbol' package has changed quite a bit since it was `site', and
the existing bash completion script is from back then.  It has been
updated to work with the current implementation.

Furthermore, there was a limitation with the completion where simbol
arguments that were options (-D or --debug for example) caused it to
break; this has also been fixed.