New functionality for `module/util
The `join' function now can take an optional third variable as an
index into the array from where to start.

A new `zip' function now combines two arrays into a single associative
array; this is an `.eval()' style function (i.e., the output of this
function be fed to `eval'.